requirements.txt
Flask==2.3.3
flask-cors==4.0.0
textblob==0.18.0
fpdf2>=2.5.6

Pillow==10.2.0
easyocr==1.7.1

python-Levenshtein==0.23.0

torch==2.2.2
opencv-python-headless==4.9.0.80
celery
numpy<2.0

celery[redis]==5.3.4
redis>=4.0.0
kombu==5.3.4

pytesseract==0.3.10

yt-dlp==2024.4.9
fuzzywuzzy==0.18.0